requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
importlib-metadata==4.10.0
Markdown==3.3.6
mimesis==5.2.1
pytz==2021.3
sqlparse==0.4.2
zipp==3.6.0
